Rename package from updateuser to update
